finish hopefully-correct impl of user accounts

+++ server.py
@@ -49,6 +49,7 @@ conn = sqlite3.connect(os.path.join(DATA_DIR, 'p.db'), check_same_thread=False)
 c = conn.cursor()
 c.executescript('''
 CREATE TABLE IF NOT EXISTS d (
+    uid     INTEGER NOT NULL,
     genre   TEXT NOT NULL,
     variant TEXT,
     flags   TEXT,
@@ -98,21 +99,21 @@ class PuzzlinkHelper(http.server.SimpleHTTPRequestHandler):
     def do_GET(self):
         self.nohtml('p')
         self.nohtml('db')
-        print(self.path)
+        self.nohtml('auth')
         super().do_GET()
 
     def do_POST(self):
         uid = c.execute('SELECT uid FROM tokens WHERE token = ?', (self.headers.get('PzplusAuth', ''),)).fetchone()
-        if not uid:
+        if not uid and self.path != '/auth':
             self.send_response(403)
             self.end_headers()
             return
 
         ret = None
         if hasattr(API, 'b_' + self.path[1:]):
-            ret = getattr(API, 'b_' + self.path[1:])(self.rfile.read(int(self.headers['Content-Length'])))
+            ret = getattr(API, 'b_' + self.path[1:])(uid, self.rfile.read(int(self.headers['Content-Length'])))
         elif hasattr(API, 'j_' + self.path[1:]):
-            ret = getattr(API, 'j_' + self.path[1:])(json.loads(self.rfile.read(int(self.headers['Content-Length']))))
+            ret = getattr(API, 'j_' + self.path[1:])(uid, json.loads(self.rfile.read(int(self.headers['Content-Length']))))
 
         if ret is not None:
             self.send_response(200)
@@ -125,7 +126,26 @@ class PuzzlinkHelper(http.server.SimpleHTTPRequestHandler):
 
 class API:
 
-    def b_localdb(data):
+    def j_auth(_, data):
+        if data.get('action') == 'login':
+            res = c.execute('SELECT rowid, pass, salt FROM users WHERE name = ?', (data.get('name'),)).fetchone()
+            if not res: return {'msg': 'no such user'}
+            uid, pwd, salt = res
+            if 'pass' not in data or pwhash(data['pass'].encode(), salt) != pwd: return {'msg': 'incorrect password'}
+            with clock: return {'msg': 'success!', 'token': maketoken(uid)}
+
+        if data.get('action') == 'register':
+            if 'name' not in data or not data['name'] or 'pass' not in data or not data['pass']: return {'msg': 'missing login information'}
+            with clock:
+                if c.execute('SELECT COUNT(*) FROM users WHERE name = ?', (data.get('name'),)).fetchone()[0]: return {'msg': 'username already taken'}
+                salt = os.urandom(32)
+                c.execute('INSERT INTO users (name, pass, salt) VALUES (?, ?, ?)', (data['name'], pwhash(data['pass'].encode(), salt), salt))
+                conn.commit()
+                return {'msg': 'success!', 'token': maketoken(c.lastrowid)}
+
+        return {'msg': 'something weird happened'}
+
+    def b_localdb(uid, data):
         data, *recording = data.split(b'\0', 1)
         data = json.loads(data)
 
@@ -133,8 +153,8 @@ class API:
         genre = patch(parts[0])
         flags, w, h = ([None]+parts[1:3]) if parts[1].isdigit() else parts[1:4]
         with clock:
-            c.execute('INSERT INTO d (genre,flags,url,date,w,h,t) VALUES (?,?,?,datetime("now","localtime"),?,?,?)',
-                    (genre, flags, data['url'], w, h, data['t']))
+            c.execute('INSERT INTO d (uid,genre,flags,url,date,w,h,t) VALUES (?,?,?,?,datetime("now","localtime"),?,?,?)',
+                    (uid, genre, flags, data['url'], w, h, data['t']))
             conn.commit()
             rowid = c.lastrowid
 
@@ -142,20 +162,20 @@ class API:
             with open(recpath(rowid), 'wb') as recfile:
                 recfile.write(recording[0])
 
-        num, time = c.execute('SELECT COUNT(*), SUM(t) FROM d WHERE genre = ?', (genre,)).fetchone()
+        num, time = c.execute('SELECT COUNT(*), SUM(t) FROM d WHERE uid = ? AND genre = ?', (uid, genre)).fetchone()
         return {
             'msg1': f'saved time: {tts(data["t"], True)}',
             'msg2': f'{num} {genre} puzzles solved in {tts(time)}',
             'rowid': rowid
         }
 
-    def j_fetch(data):
+    def j_fetch(uid, data):
         # oops, bunch of copy/paste from above
         parts = data['url'].split('/')
         genre = patch(parts[0])
-        num, time = c.execute('SELECT COUNT(*), SUM(t) FROM d WHERE genre = ?', (genre,)).fetchone()
+        num, time = c.execute('SELECT COUNT(*), SUM(t) FROM d WHERE uid = ? AND genre = ?', (uid, genre)).fetchone()
         rowid, t, rate, diff, path, uniq, comm, variant = \
-            c.execute('SELECT rowid, t, rate, diff, path, uniq, comm, variant FROM d WHERE url = ?', (data['url'],)).fetchone()
+            c.execute('SELECT rowid, t, rate, diff, path, uniq, comm, variant FROM d WHERE uid = ? AND url = ?', (uid, data['url'])).fetchone()
         return {
             'msg1': f'editing existing time: {tts(t, True)}',
             'msg2': f'{num} {genre} puzzles solved in {tts(time)}',
@@ -163,29 +183,31 @@ class API:
             'rate': rate, 'diff': diff, 'path': path, 'uniq': uniq, 'comm': comm, 'variant': variant
         }
 
-    def j_update(data):
+    def j_update(uid, data):
+        if c.execute('SELECT COUNT(*) FROM d WHERE uid = ? AND rowid = ?', (uid, data.get('rowid'))).fetchone()[0] == 0:
+            return {'msg': 'no such solve'}
         if data['k'] in ['rate', 'diff', 'path', 'uniq', 'variant', 'comm']:
             with clock:
-                c.execute(f'UPDATE d SET {data["k"]} = ? WHERE rowid = ?', (data['v'], data['rowid']))
+                c.execute(f'UPDATE d SET {data["k"]} = ? WHERE uid = ? AND rowid = ?', (data['v'], uid, data['rowid']))
                 conn.commit()
             return { 'msg': 'saved!' }
         if data['k'] == 'unsave' and data['v'] == 1:
             with clock:
-                c.execute('DELETE FROM d WHERE rowid = ?', (data['rowid'],))
+                c.execute('DELETE FROM d WHERE uid = ? AND rowid = ?', (uid, data['rowid']))
                 conn.commit()
             try: os.remove(recpath(data['rowid']))
             except: pass
             return { 'msg': 'deleted' }
         return { 'msg': 'error' }
 
-    def j_getrec(data):
-        res = c.execute('SELECT rowid FROM d WHERE url = ?', (data['url'],)).fetchone()
+    def j_getrec(uid, data):
+        res = c.execute('SELECT rowid FROM d WHERE uid = ? AND url = ?', (uid, data['url'])).fetchone()
         fname = recpath(res[0]) if res else None
         return open(fname, 'rb').read() if fname else b''
 
-    def j_prevsolves(data):
+    def j_prevsolves(uid, data):
         return [{
             't': tts(t)
-        } for (t,) in c.execute('SELECT t FROM d WHERE url = ?', (data['url'],)).fetchall()]
+        } for (t,) in c.execute('SELECT t FROM d WHERE uid = ? AND url = ?', (uid, data['url'])).fetchall()]
 
 http.server.ThreadingHTTPServer(('', PORT), PuzzlinkHelper).serve_forever()
